+++ src/Components/Modal/ChangeEmploymentModal/index.tsx
@@ -4,7 +4,7 @@ import * as _ from '../style';
 import { useForm } from '../../../Hooks/useForm';
 import { Pagination } from '../../../Utils/Pagination';
 import { useChangeEmployment } from '../../../Apis/Acceptances';
-import { useGetCode } from '../../../Hooks/ApiHooks/useGetCode';
+import { useGetCode } from '../../../Hooks/ApiHooks/Codes';
 
 interface PropType {
 	selectStudent: number[];
