+++ src/Hooks/ApiHooks/Acceptances.tsx
@@ -1,5 +1,5 @@
-import { useQuery } from "react-query";
-import { getCombinedStudentList } from "../../Apis/Acceptances";
+import { useQuery } from 'react-query';
+import { getCombinedStudentList } from '../../Apis/Acceptances';
 
 /** 현장실습 학생과 근로계약 학생을 조회하는 api입니다. */
 export function useGetCombinedStudentList(company_id: number) {
