+++ src/Pages/CompanyRecruitmentPage/index.tsx
@@ -4,7 +4,7 @@ import { CompanyRecruitmentTable } from '../../Components/CompanyRecruitment/Tab
 import { Header } from '../../Components/Header';
 import { useState } from 'react';
 import { dataType } from '../../Apis/Companies/request';
-import { useGetCompanyRecruitments } from '../../Hooks/ApiHooks/useGetCompanyRecruitments';
+import { useGetCompanyRecruitments } from '../../Hooks/ApiHooks/Companies';
 
 export function CompanyRecruitmentPage() {
 	const [searchQueryString, setSearchQueryString] = useState<dataType>({
