+++ src/Pages/RecruitmentFormPage/index.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
-import { useGetRecruitmentForm } from '../../Hooks/ApiHooks/useGetRecruitments';
 import * as _ from './style';
 import { RecruitmentFormQueryStringType } from '../../Apis/Recruitments/request';
 import { Header } from '../../Components/Header';
 import { RecruitmentFormSearch } from '../../Components/RecruitmentForm/Search';
 import { RecruitmentFormTable } from '../../Components/RecruitmentForm/Table';
+import { useGetRecruitmentForm } from '../../Hooks/ApiHooks/Applications';
 
 export function RecruitmentFormPage() {
 	const date = new Date(); // 현재 날짜 및 시간
